+++ package.json
@@ -16,6 +16,7 @@
     "@types/react-dom": "^16.9.5",
     "@types/react-redux": "^7.1.7",
     "axios": "^0.19.2",
+    "axios-mock-adapter": "^1.17.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-redux": "^7.2.0",
